HDDS-2184. Rename ozone scmcli to ozone admin
