+++ hadoop-hdds/tools/src/main/java/org/apache/hadoop/hdds/scm/cli/pipeline/PipelineCommands.java
@@ -17,13 +17,15 @@
  */
 package org.apache.hadoop.hdds.scm.cli.pipeline;
 
+import java.util.concurrent.Callable;
+
 import org.apache.hadoop.hdds.cli.HddsVersionProvider;
 import org.apache.hadoop.hdds.cli.MissingSubcommandException;
-import org.apache.hadoop.hdds.scm.cli.SCMCLI;
+import org.apache.hadoop.hdds.scm.cli.container.WithScmClient;
+
+import picocli.CommandLine;
 import picocli.CommandLine.Command;
 import picocli.CommandLine.ParentCommand;
-
-import java.util.concurrent.Callable;
 
 /**
  * Subcommand to group pipeline related operations.
@@ -43,15 +45,15 @@ import java.util.concurrent.Callable;
 public class PipelineCommands implements Callable<Void> {
 
   @ParentCommand
-  private SCMCLI parent;
+  private WithScmClient parent;
 
-  public SCMCLI getParent() {
+  public WithScmClient getParent() {
     return parent;
   }
 
   @Override
   public Void call() throws Exception {
     throw new MissingSubcommandException(
-        this.parent.getCmd().getSubcommands().get("pipeline"));
+        new CommandLine(new PipelineCommands()));
   }
 }
